#42 refactored the mapping logic in BeanMapper to separate strategies for mapping to collections, constructable classes or instances. The strategy to execute is determined by the configuration; i) if a collection class is defined, ii) if a target class is defined, or iii) if a target instance is defined.
NumberToNumberConverter called convert() on BeanMapper. It now calls the top-level map method, including a converterChoosable bit to make sure it will select the default converters.
