+++ src/main/java/io/beanmapper/strategy/MapStrategy.java
@@ -1,0 +1,7 @@
+package io.beanmapper.strategy;
+
+public interface MapStrategy {
+
+    Object map(Object source);
+
+}
